Buncha big changes
Created the DFormerComponent abstract class which deformers will derive
from.

Renamed DFormObjectBase to DFormManagerBase.
Created DFormManager which inherits from DFormMangerBase and works
similarly to the ModifierObject in my other Mesh Modifier package (it it
component-based)

Created a sample deformer to test that just adds an offset to the mesh.

Works in the editor!

Created MeshUtil class. Right now all it does is make copying meshes
easier.

Chunk's "Size" is now a property so cache it if you are using it alot.