Big fixes
Fixed bug where duplicating an object would delete the shared mesh on
the original object.

Had to revert the changes that let deformers continue running when not
selected because on destroy wasn't called when an object was deleted via
undoing its creations, so the editor callback that told the object to
update would reference a null object and throw errors for eternity.